requirements.txt
asgiref==3.4.1
astroid==2.9.0
attrs==21.4.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0
django-import-export==2.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==10.0.0
filelock==3.4.2
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.3.2
lazy-object-proxy==1.7.1
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.1
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
tablib==3.1.0
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
